BinBot_Capstone_project/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pillow
python-multipart
tensorflow